requirements.txt
# requirements.txt for tinybwmon
Flask
speedtest-cli
